execution/scrape_single_site.py
import os
import sys
import json
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def scrape_site(url):
    """
    Deterministic execution script for scraping.
    In a real scenario, this would use requests or firecrawl.
    """
    print(f"Scraping {url}...")
    
    # Mock data processing
    data = {
        "url": url,
        "title": "Example Domain",
        "features": ["Feature A", "Feature B"]
    }
    
    # Ensure .tmp directory exists
    os.makedirs(".tmp", exist_ok=True)
    
    # Save to intermediate file
    output_path = os.path.join(".tmp", "scraped_data.json")
    with open(output_path, "w") as f:
        json.dump(data, f, indent=2)
        
    print(f"Results saved to {output_path}")

if __name__ == "__main__":
    if len(sys.argv) < 2:
        print("Usage: python scrape_single_site.py <url>")
        sys.exit(1)
        
    target_url = sys.argv[1]
    scrape_site(target_url)
创新
