requirements.txt
requests
python-dotenv
# beautifulsoup4
# firecrawl-py
